requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.6
django-cors-headers==3.13.0
django-countries==7.3.2
django-filter==22.1
dj-rest-auth==2.2.5
dj-rest-auth[with_social]==2.2.5
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.0
idna==3.3
install==1.3.5
psycopg2==2.9.3
PyJWT==1.7.1
pytz==2022.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.3.0
urllib3==1.26.11
